Refactored course routes to replace `$courseId` with `$courseCode`, removed legacy components and logic, and integrated streamlined `CourseNavBar` for improved maintainability and consistency.
